+++ example/src/hooks/useRunTests.ts
@@ -1,7 +1,9 @@
+import 'mocha';
+import type * as MochaTypes from 'mocha';
 import { useCallback, useEffect, useState } from 'react';
 import type { TestItemType } from '../navigators/children/Entry/TestItemType';
 import type { RowItemType } from '../navigators/children/TestingScreen/RowItemType';
-import { testLib } from '../testing/MochaSetup';
+import { clearTests, rootSuite } from '../testing/MochaRNAdapter';
 
 export const useRunTests = (
   tests: TestItemType[]
@@ -21,7 +23,7 @@ export const useRunTests = (
       const testRegistrators: (() => void)[] = tests
         .filter((t) => t.value)
         .map((t) => t.registrator);
-      testLib(addResult, testRegistrators);
+      run(addResult, testRegistrators);
     },
     // eslint-disable-next-line react-hooks/exhaustive-deps
     [results]
@@ -33,4 +35,75 @@ export const useRunTests = (
 
   console.log({ results });
   return [results, runTests];
+};
+
+const run = (
+  addTestResult: (testResult: RowItemType) => void,
+  testRegistrators: Array<() => void> = []
+) => {
+  // console.log('setting up mocha');
+
+  const {
+    EVENT_RUN_BEGIN,
+    EVENT_RUN_END,
+    EVENT_TEST_FAIL,
+    EVENT_TEST_PASS,
+    EVENT_SUITE_BEGIN,
+    EVENT_SUITE_END,
+  } = Mocha.Runner.constants;
+
+  clearTests();
+  var runner = new Mocha.Runner(rootSuite) as MochaTypes.Runner;
+
+  let indents = -1;
+  const indent = () => Array(indents).join('  ');
+  runner
+    .once(EVENT_RUN_BEGIN, () => {})
+    .on(EVENT_SUITE_BEGIN, (suite: MochaTypes.Suite) => {
+      const name = suite.fullTitle();
+      if (name !== '') {
+        addTestResult({
+          indentation: indents,
+          description: name,
+          key: Math.random().toString(),
+          type: 'grouping',
+        });
+      }
+      indents++;
+    })
+    .on(EVENT_SUITE_END, () => {
+      indents--;
+    })
+    .on(EVENT_TEST_PASS, (test: MochaTypes.Runnable) => {
+      addTestResult({
+        indentation: indents,
+        description: test.fullTitle(),
+        key: Math.random().toString(),
+        type: 'correct',
+      });
+      console.log(`${indent()}pass: ${test.fullTitle()}`);
+    })
+    .on(EVENT_TEST_FAIL, (test: MochaTypes.Runnable, err: Error) => {
+      addTestResult({
+        indentation: indents,
+        description: test.fullTitle(),
+        key: Math.random().toString(),
+        type: 'incorrect',
+        errorMsg: err.message,
+      });
+      console.error(
+        `${indent()}fail: ${test.fullTitle()} - error: ${err.message}`
+      );
+    })
+    .once(EVENT_RUN_END, () => {});
+
+  testRegistrators.map((register) => {
+    register();
+  });
+  runner.run();
+
+  return () => {
+    console.log('aborting');
+    runner.abort();
+  };
 };
