+++ example/src/testing/Tests/pbkdf2Tests/pbkdf2Tests.ts
@@ -9,7 +9,7 @@ function ab2str(buf: ArrayBuffer) {
   return Buffer.from(buf).toString('hex');
 }
 
-export const pbkdf2RegisterTests = () => {
+export const registerPbkdf2Tests = () => {
   // Copied from https://github.com/crypto-browserify/pbkdf2/blob/master/test/index.js
   // SHA-1 vectors generated by Node.js
   // SHA-256/SHA-512 test vectors from:
@@ -19,7 +19,7 @@ export const pbkdf2RegisterTests = () => {
   // eslint-disable-next-line @typescript-eslint/no-shadow
   var Buffer = require('safe-buffer').Buffer;
 
-  describe('pbkdf2 tests', () => {
+  describe('pbkdf2', () => {
     it(' defaults to sha1 and handles buffers', (done: Done) => {
       var resultSync = QuickCrypto.pbkdf2Sync('password', 'salt', 1, 32);
       chai
